test(bootstrap): split GitHub CLI auth tests into their own file
test_bootstrap.py crossed the 500-line taste-lint threshold once
TestConfigureGithubCli and TestRestoreOriginRemote were added. Extract
both into test_bootstrap_github_cli.py; no behavior change.

Refs #5277
